Move .y4m parser to a test-only file.
The external API currently doesn't support encoding subsampled YUV
frames, so we can't support this mode in cjxl. This patch moves the .y4m
parsing to another test-only file that is used by jxl_test.cc so we can
continue to run the roundtrip test (which upsamples the data). This
removes the possibility to encode .y4m images in the cjxl tool.
