+++ lib/jxl/codec_y4m_testonly.h
@@ -1,0 +1,18 @@
+// Copyright (c) the JPEG XL Project Authors. All rights reserved.
+//
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+#include <stdint.h>
+
+#include "lib/jxl/base/padded_bytes.h"
+#include "lib/jxl/base/status.h"
+#include "lib/jxl/codec_in_out.h"
+
+namespace jxl {
+namespace test {
+
+Status DecodeImageY4M(const Span<const uint8_t> bytes, CodecInOut* io);
+
+}  // namespace test
+}  // namespace jxl
