+++ lib/lib.gni
@@ -383,6 +383,8 @@ libjxl_tests_sources = [
 
 # Test-only library code.
 libjxl_testlib_sources = [
+    "jxl/codec_y4m_testonly.cc",
+    "jxl/codec_y4m_testonly.h",
     "jxl/dct_for_test.h",
     "jxl/dec_transforms_testonly.cc",
     "jxl/dec_transforms_testonly.h",
